+++ hardware/include/diffdrive_esp32/wheel.hpp
@@ -10,7 +10,7 @@ class Wheel
     public:
 
     std::string name = "";
-    int enc = 0;
+    int64_t enc = 0;
     double cmd = 0;
     double pos = 0;
     double vel = 0;
